requirements.txt
flask==1.1.2
Pillow==8.2.0
pylast==4.2.0
requests==2.25.1
